KOCO-bench-en/domain_code_generation/raganything/test_examples/BookWorm/code/pyproject.toml
[project]
name = "bookworm"
version = "0.1.0"
description = "Advanced document/knowledge ingestion system with LightRAG and mindmap generation"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # PDF processing (following user's mineru methodology)
    "mineru>=0.2.2", # Primary PDF processor like in lightrag_ex.py
    "pymupdf>=1.23.0", # Fallback PDF processor
    "pdfplumber>=0.9.0", # Alternative PDF processor
    # Document processing
    "python-docx>=0.8.11",
    "openpyxl>=3.1.0",
    "python-pptx>=0.6.21",
    # LLM integrations (matching user's setup)
    "openai>=1.0.0",
    "anthropic>=0.7.0",
    "google-generativeai>=0.3.0",
    # Mindmap generation dependencies
    "termcolor>=2.0.0",
    "aiofiles>=23.2.1",
    "fuzzywuzzy>=0.18.0",
    "python-decouple>=3.8",
    # Utilities
    "pydantic>=2.5.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "loguru>=0.7.0",
    "httpx>=0.25.0",
    "tqdm>=4.66.0",
    # Vector storage (for LightRAG)
    "faiss-cpu>=1.7.4",
    "numpy>=1.24.0",
    # Environment management
    "python-dotenv>=1.0.0",
    "ollama>=0.5.3",
    "uvloop>=0.21.0",
    "lightrag-hku>=1.4.6",
    "raganything[all]>=1.2.6",
    "update>=0.0.1",
    "chartographer",
    "python-levenshtein>=0.27.1",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "black",
    "isort",
    "mypy",
]

[project.scripts]
bookworm = "bookworm.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"

[tool.uv.sources]
chartographer = { git = "https://github.com/haxx0rman/Chartographer" }
